delete posts integeration fixes
